+++ packages/jest-config/src/__tests__/normalize.test.js
@@ -1026,3 +1026,73 @@ describe('watchPlugins', () => {
     ]);
   });
 });
+
+describe('testPathPattern', () => {
+  const initialOptions = {rootDir: '/root'};
+  const consoleLog = console.log;
+
+  function testWindowsPathSeparator(argv, expected) {
+    jest.resetModules();
+    jest.mock('path', () => require.requireActual('path').win32);
+    require('jest-resolve').findNodeModule = findNodeModule;
+
+    const {options} = require('../normalize').default(initialOptions, argv);
+    expect(options.testPathPattern).toBe(expected);
+  }
+
+  beforeEach(() => {
+    console.log = jest.fn();
+  });
+
+  afterEach(() => {
+    console.log = consoleLog;
+  });
+
+  it('defaults to empty', () => {
+    const {options} = normalize(initialOptions, {});
+    expect(options.testPathPattern).toBe('');
+  });
+
+  describe('--testPathPattern', () => {
+    it('uses testPathPattern if set', () => {
+      const {options} = normalize(initialOptions, {testPathPattern: 'a/b'});
+      expect(options.testPathPattern).toBe('a/b');
+    });
+
+    it('ignores invalid regular expressions and logs a warning', () => {
+      const {options} = normalize(initialOptions, {testPathPattern: 'a('});
+      expect(options.testPathPattern).toBe('');
+      expect(console.log.mock.calls[0][0]).toMatchSnapshot();
+    });
+
+    it('escapes Windows path separators', () => {
+      testWindowsPathSeparator({testPathPattern: 'a\\b'}, 'a\\\\b');
+    });
+  });
+
+  describe('<regexForTestFiles>', () => {
+    it('uses <regexForTestFiles> if set', () => {
+      const {options} = normalize(initialOptions, {_: ['a/b']});
+      expect(options.testPathPattern).toBe('a/b');
+    });
+
+    it('ignores invalid regular expressions and logs a warning', () => {
+      const {options} = normalize(initialOptions, {_: ['a(']});
+      expect(options.testPathPattern).toBe('');
+      expect(console.log.mock.calls[0][0]).toMatchSnapshot();
+    });
+
+    it('escapes Windows path separators', () => {
+      testWindowsPathSeparator({_: ['a\\b']}, 'a\\\\b');
+    });
+
+    it('joins multiple <regexForTestFiles> if set', () => {
+      const {options} = normalize(initialOptions, {_: ['a/b', 'c/d']});
+      expect(options.testPathPattern).toBe('a/b|c/d');
+    });
+
+    it('escapes Windows path separators in multiple args', () => {
+      testWindowsPathSeparator({_: ['a\\b', 'c\\d']}, 'a\\\\b|c\\\\d');
+    });
+  });
+});